backend/requirements.txt
# Web
fastapi==0.115.4
uvicorn[standard]==0.30.6
websockets>=12.0,<13.0

# HTTP / cliente
httpx>=0.27.2,<0.28       # cliente HTTP para llamadas REST a Supabase y otros servicios
python-multipart==0.0.9   # subir ficheros
aiofiles==24.1.0

# Config / utilidades
python-dotenv==1.0.1
loguru==0.7.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2              # <- fija versión compatible y evita warning/errores

# DB
psycopg2-binary==2.9.9
# Nota: Ya no usamos supabase-py, usamos httpx directamente para llamadas REST

# Tipado / validación
pydantic==2.9.2
pydantic-settings==2.7.0

# Crypto
cryptography==43.0.3

# PDF Generation
reportlab==4.0.9

# Google AI
google-generativeai>=0.7.0,<0.8.0

# OpenAI (fallback para rate limits)
openai==1.58.1

# Google APIs (Drive)
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# SSL certificates
certifi==2024.12.14

# Web automation
playwright>=1.48.0

# LangChain & LangGraph for agent workflows
# Versiones compatibles: langchain-core 0.3.x con langgraph 0.2.x (más reciente)
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-openai==0.2.0
langgraph==0.2.40

# Web search for online capabilities
duckduckgo-search==6.1.0

# Async support for tools
nest-asyncio==1.6.0

# Email service (Resend)
resend==2.0.0
